Add tests. Add array
Former-commit-id: 84d379f46e255631b8d2c140b738ce63b01d93a7

+++ include/kb/handle.h
@@ -8,7 +8,7 @@ extern "C" {
 
 typedef uint32_t Handle;
 
-static const Handle InvalidHandle = UINT16_MAX;
+static const Handle InvalidHandle = UINT32_MAX;
 
 static inline bool is_valid_handle(const Handle handle) { return handle != InvalidHandle; }
